requirements.txt
llama-index==0.10.0
neo4j==5.11.0
python-dotenv==1.0.0
openai==1.3.6
